+++ NEWS.md
@@ -71,7 +71,7 @@ reduce2_right(.x = letters[1:4], .y = paste2, .f = c("-", ".", "-")) # working
 * New `assign_in()` and `pluck<-` functions. They modify a data
   structure at a pluck location.
 
-* New `pluck_modify()` function to map a function at a pluck location.
+* New `modify_in()` function to map a function at a pluck location.
 
 * Functions taking predicates (`map_if()`, `keep()`, `some()`,
   `every()`, `keep()`, etc) now fail with an informative message when